include CNAME in build process
I noticed that a change to CNAME was made directly to docs folder,
since we're manually generating the `docs/` and remove it on
`make clean` the changes to the CNAME won't persist.

So the root `/CNAME` file is the canonical file now, which will be copied
to `/docs/CNAME` on every build.