Make the event link (now join) stand out from the other links
I thought that moving it to the last column would be better because,
as the website is in English (read left to right), the "registration"
or "join" invitation fits more naturally at the end of a row of
information.

After trying it out though, I realized that the position at the
beginning, maybe exactly because it inverts the assumed order, helps
highlight the importance of the registration process, which is the
main goal here.