Improve navigation, move ns from partials to components
The previous navigation was part of the initial design and, with the
growing number of links, was completely broken on mobile and not
looking good on desktops.

I have also renamed `views/partials.clj` to
`views/components/navigation.clj` because this is closer to the
convention I have used later in the project.